# docs/swagger.yaml
basePath: /api
definitions:
...
  lab_internal_app_DTO.ChangeUserData:
    properties:
      login:
        type: string
    type: object
  lab_internal_app_DTO.LoginReq:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  lab_internal_app_DTO.LoginRes:
    properties:
      access_token:
        type: string
      expires_in:
        $ref: '#/definitions/time.Duration'
      is_moderator:
        type: boolean
      token_type:
        type: string
    type: object
...
  lab_internal_app_DTO.UserDataResposne:
    properties:
      id:
        type: integer
      login:
        type: string
    type: object
  lab_internal_app_DTO.UserRegistration:
    properties:
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
...
  time.Duration:
    enum:
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    format: int64
    type: integer
    x-enum-varnames:
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
host: localhost:8001
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API для управления солнечными панелями и расчета их мощности
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Solar Panels Power Calculator API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Выполняет вход в систему и возвращает JWT токен
      parameters:
      - description: Логин и пароль
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/lab_internal_app_DTO.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: JWT токен и информация о сессии
          schema:
            $ref: '#/definitions/lab_internal_app_DTO.LoginRes'
        "400":
          description: Некорректные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Неверный логин или пароль
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Авторизация пользователя
      tags:
      - Users
  /logout:
    post:
      consumes:
      - application/json
      description: Добавляет JWT токен в черный список
      produces:
      - application/json
      responses:
        "200":
          description: Успешный выход
        "400":
          description: Некорректный токен
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Выход из системы
      tags:
      - Users
...
  /user:
    put:
      consumes:
      - application/json
      description: Обновляет данные текущего авторизованного пользователя
      parameters:
      - description: Обновленные данные пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/lab_internal_app_DTO.ChangeUserData'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленные данные пользователя
          schema:
            $ref: '#/definitions/lab_internal_app_DTO.UserDataResposne'
        "400":
          description: Некорректные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Изменить данные пользователя
      tags:
      - Users
  /user/{id}:
    get:
      consumes:
      - application/json
      description: Возвращает информацию о пользователе по ID
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Данные пользователя
          schema:
            $ref: '#/definitions/lab_internal_app_DTO.UserDataResposne'
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Пользователь не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить данные пользователя
      tags:
      - Users
  /user/registration:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя в системе
      parameters:
      - description: Данные для регистрации
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/lab_internal_app_DTO.UserRegistration'
      produces:
      - application/json
      responses:
        "201":
          description: Данные созданного пользователя
          schema:
            $ref: '#/definitions/lab_internal_app_DTO.UserDataResposne'
        "400":
          description: Некорректные данные или пользователь уже существует
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация пользователя
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"